i32

+++ src/esp32/encoder.rs
@@ -199,7 +199,7 @@ where
             }
         }
 
-        isr_install(self.config.unit)?;
+        isr_install(self.config.unit as i32)?;
 
         esp!(unsafe {
             esp_idf_sys::pcnt_isr_handler_add(
